chore: cleanup tests
