+++ tests/transcription_evals/test_metrics.py
@@ -11,13 +11,13 @@ def test_normalise_text_edge_cases():
 
 
 def test_compute_wer_metrics_empty_reference_raises_error():
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match=".*"):
         compute_wer_metrics([""], ["hello world"])
 
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match=".*"):
         compute_wer_metrics([""], [""])
 
-    with pytest.raises(ValueError):
+    with pytest.raises(ValueError, match=".*"):
         compute_wer_metrics([], [])
